requirements.txt
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
PyQt5>=5.15.0
scikit-image>=0.21.0
onnxruntime>=1.15.0
requests>=2.31.0